jwt service add

+++ service/map.service.js
@@ -1,4 +1,5 @@
 const map = require("../repository/map.repository.js");
+const data = require("../utils/getData_utils");
 
 exports.applyMap = (req, res) => {
   if (!req.body) {
@@ -70,28 +71,32 @@ exports.updateMapInfo = (req, res) => {
   }
 };
 
-exports.deleteMap = (req, res) => {
+exports.deleteMap = async (req, res) => {
   if (!req.params) {
     res.status(400).send({
       message: "req.params can not be empty!",
     });
-  }
-//   else if (!req.user) {
-//     res.status(401).send({
-//       message: "Unauthorized",
-//     });
-//   }
-  else {
-    map.deleteMap(req.params.map_id, (err, results) => {
-      if (err) {
-        res.status(400).send({
-          message: "deleteMap err",
-        });
-      } else {
-        res.send({
-          message: `map(${req.params.map_id}) was successfully deleted`,
-        });
-      }
-    });
+  } else {
+    const forDelete = await data.findMapById(req.params.map_id)
+    if (forDelete.ok === false) {
+      res.status(400).send({ message: `There is no map(${req.params.map_id})` });
+    }
+    else if (req.user !== forDelete.map.map_maker) {
+      res.status(401).send({
+          message: "Unauthorized. Can only delete my map."
+      });
+    } else {
+      map.deleteMap(req.params.map_id, (err, results) => {
+        if (err) {
+          res.status(400).send({
+            message: "deleteMap err",
+          });
+        } else {
+          res.send({
+            message: `map(${req.params.map_id}) was successfully deleted`,
+          });
+        }
+      });
+    }
   }
 };
